+++ src/rich_click/patch.py
@@ -6,6 +6,7 @@ from typing import Optional
 
 import click
 
+import rich_click
 import rich_click.rich_command
 from rich_click.decorators import command as _rich_command
 from rich_click.decorators import group as _rich_group
@@ -39,7 +40,7 @@ def rich_group(*args, **kwargs):  # type: ignore[no-untyped-def]
     return _rich_group(*args, **kwargs)
 
 
-def patch(rich_config: Optional[RichHelpConfiguration] = None) -> None:
+def patch(rich_config: Optional[RichHelpConfiguration] = None, patch_rich_click: bool = True) -> None:
     """Patch Click internals to use rich-click types."""
     from rich_click._compat_click import CLICK_IS_BEFORE_VERSION_9X
 
@@ -51,5 +52,13 @@ def patch(rich_config: Optional[RichHelpConfiguration] = None) -> None:
     click.CommandCollection = _PatchedRichCommandCollection  # type: ignore[misc]
     if CLICK_IS_BEFORE_VERSION_9X:
         click.MultiCommand = _PatchedRichMultiCommand  # type: ignore[assignment,misc,unused-ignore]
+    if patch_rich_click:
+        rich_click.group = rich_group
+        rich_click.command = rich_command
+        rich_click.Group = _PatchedRichGroup  # type: ignore[misc]
+        rich_click.Command = _PatchedRichCommand  # type: ignore[misc]
+        rich_click.CommandCollection = _PatchedRichCommandCollection  # type: ignore[misc]
+        if CLICK_IS_BEFORE_VERSION_9X:
+            rich_click.MultiCommand = _PatchedRichMultiCommand  # type: ignore[assignment,misc,unused-ignore]
     if rich_config is not None:
         rich_config.dump_to_globals()